+++ crates/pacdef/src/errors.rs
@@ -6,8 +6,6 @@ use std::path::PathBuf;
 #[derive(Debug)]
 #[non_exhaustive]
 pub enum Error {
-    /// Package search yields no results.
-    NoPackagesFound,
     /// Config file not found.
     ConfigFileNotFound,
     /// Group file not found.
@@ -23,7 +21,6 @@ pub enum Error {
 impl Display for Error {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            Self::NoPackagesFound => write!(f, "no packages matching query"),
             Self::ConfigFileNotFound => write!(f, "config file not found"),
             Self::GroupFileNotFound(name) => write!(f, "group file '{name}' not found"),
             Self::GroupAlreadyExists(path) => {
