+++ crates/pacdef/src/prelude.rs
@@ -12,13 +12,13 @@ pub use crate::cli::CleanPackageAction;
 pub use crate::cli::MainArguments;
 pub use crate::cli::MainSubcommand;
 pub use crate::cli::ReviewPackageAction;
-pub use crate::cli::SearchPackageAction;
 pub use crate::cli::SyncPackageAction;
 pub use crate::cli::UnmanagedPackageAction;
 pub use crate::cli::VersionArguments;
 pub use crate::config::Config;
 pub use crate::errors::Error;
-pub use crate::packages::{Groups, PackagesIds, PackagesInstall, PackagesQuery, PackagesRemove};
+pub use crate::groups::Groups;
+pub use crate::packages::{PackagesIds, PackagesInstall, PackagesQuery, PackagesRemove};
 pub use crate::path::binary_in_path;
 pub use crate::path::get_absolutized_file_paths;
 pub use crate::path::get_cargo_home;
