+++ man/pacdef.8
@@ -153,7 +153,7 @@ Show version information (including git revision if it was build from git) and s
 
 
 .SH EXIT STATUS
-Pacdef exits with status 0 on success, 1 if an error occurs (e.g. package search did not yield any package), and 2 if invalid command line options were specified.
+Pacdef exits with status 0 on success, 1 if an error occurs, and 2 if invalid command line options were specified.
 
 .SH BUGS
 File bugs and feature requests under https://github.com/steven-omaha/pacdef/issues.
